password_manager/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
cryptography
passlib[bcrypt]
python-jose
PyQt5
requests
